[Front] migration of Groupings and GroupingsLines to tsx (#1809)
